Node implementations use generics
Altered implementations of Node to use generics in their implementations
of the evaluate method.

+++ src/test/java/org/oakgp/examples/gridwar/GridWar.java
@@ -86,7 +86,7 @@ class GridWar implements TwoPlayerGame {
    }
 
    private static int getNextMove(Player player, Assignments assignments) {
-      int result = (int) player.getLogic().evaluate(assignments);
+      int result = player.getLogic().evaluate(assignments);
       // normalise the result to ensure it is in the valid range of possible moves
       return Math.abs(result % NUMBER_OF_POSSIBLE_DIRECTIONS);
    }
